% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_DDshiny.R
\docType{data}
\name{DDchildren}
\alias{DDchildren}
\title{Shiny applet for family size example}
\format{An object of class \code{shiny.appobj} of length 5.}
\usage{
DDchildren
}
\description{
Shiny applet for family size example
}
\keyword{datasets}
