% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/namespace.R, R/zzz_DDshiny.R
\docType{data}
\name{DDshiny}
\alias{DDshiny}
\alias{DDshiny}
\title{DeclareDesign Wizard}
\format{An object of class \code{shiny.appobj} of length 5.}
\usage{
DDshiny
}
\description{
DeclareDesign Wizard

DDshiny wizard
}
\keyword{datasets}
