% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_DDshiny.R
\docType{data}
\name{DDshiny_legacy}
\alias{DDshiny_legacy}
\title{DDshiny wizard}
\format{An object of class \code{shiny.appobj} of length 5.}
\usage{
DDshiny_legacy
}
\description{
DDshiny wizard
}
\keyword{datasets}
