% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_DDshiny.R
\docType{data}
\name{DDwizard}
\alias{DDwizard}
\title{DDshiny wizard
DDwizard for simple designs}
\format{An object of class \code{shiny.appobj} of length 5.}
\usage{
DDwizard
}
\description{
DDshiny wizard
DDwizard for simple designs
}
\keyword{datasets}
